# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Edu0liver_prototype-healthy-api_internal_modules_handover_dto.ReplyRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
  github_com_Edu0liver_prototype-healthy-api_internal_modules_iam_dto.InviteRequest:
    properties:
      email:
        type: string
      name:
        type: string
      role:
        enum:
        - admin
        - operator
        - knowledge_manager
        type: string
    required:
    - email
    - role
    type: object
...
  github_com_Edu0liver_prototype-healthy-api_internal_modules_iam_dto.RegisterAdminRequest:
    properties:
      company_id:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - company_id
    - email
    - password
    type: object
...
  github_com_Edu0liver_prototype-healthy-api_internal_modules_iam_dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
...
  github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: SaaS multi-tenant white-label de atendimento automatizado e omnichannel
    (WhatsApp/Instagram) com IA + RAG.
  title: API
  version: "1.0"
paths:
...
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - users
...
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_iam_dto.RegisterAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_iam_dto.UserResponse'
        "400":
          description: invalid request body
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "409":
          description: email already in use or company already has users
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      summary: Register first admin
      tags:
      - auth
...
  /conversations/{id}/handover/reply:
    post:
      consumes:
      - application/json
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: Message
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_handover_dto.ReplyRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id or body
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: conversation not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "409":
          description: conversation is not under human control
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reply as operator
      tags:
      - handover
  /conversations/{id}/handover/return:
    post:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: conversation not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Return conversation to AI
      tags:
      - handover
...
  /documents/{docId}:
    delete:
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: document not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete document
      tags:
      - knowledge
...
  /knowledge-bases/{id}:
    delete:
      parameters:
      - description: Knowledge base ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: knowledge base not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete knowledge base
      tags:
      - knowledge
    get:
      parameters:
      - description: Knowledge base ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: knowledge base not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get knowledge base
      tags:
      - knowledge
...
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_iam_dto.UserResponse'
              type: array
            type: object
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: Invite
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_iam_dto.InviteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_iam_dto.UserResponse'
        "400":
          description: invalid role or body
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "409":
          description: email already in use
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite user
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"